requirements.txt
# Core
requests>=2.31.0
pyyaml>=6.0.0
jinja2>=3.1.0

# Web Scraping
beautifulsoup4>=4.12.0
selenium>=4.15.0
lxml>=4.9.0

# Documents
python-docx>=1.0.0
markdown>=3.5.0

# PDF
PyPDF2>=3.0.0
pdfplumber>=0.10.0
reportlab>=4.0.0

# Spreadsheets
openpyxl>=3.1.0
pandas>=2.1.0

# Desktop Automation
pyautogui>=0.9.54
pytesseract>=0.3.10
Pillow>=10.0.0

# Database
sqlalchemy>=2.0.0

# Scheduling & Monitoring
schedule>=1.2.0
watchdog>=3.0.0
